requirements.txt
streamlit>=1.28.0
folium>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.7.0
geopandas>=0.13.0
shapely>=2.0.0
streamlit-folium>=0.13.0
plotly>=5.15.0
python-dotenv>=1.0.0
